requirements.txt
PySide6==6.9.0
Pillow
